+++ versions.tf
@@ -1,4 +1,3 @@
-
 terraform {
   required_version = ">= 1.9.5"
   required_providers {
@@ -9,6 +8,10 @@ terraform {
     random = {
       source  = "hashicorp/random"
       version = "~> 3.8.1"
+    }
+    http = {
+      source  = "hashicorp/http"
+      version = "~> 3.5.0"
     }
   }
 }